chore: Refactor the common components directory
