+++ pages/home/cafe.js
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Typography from '@mui/material/Typography'
-import Section from '@/components/layout/section'
-import Navigation from '@/components/layout/navigation'
+import Section from '@/components/common/layout/section'
+import Navigation from '@/components/common/layout/navigation'
 
 import vercelLogo from '@/public/vercel.svg'
