+++ pages/index.js
@@ -1,6 +1,6 @@
 import Head from 'next/head'
 import Image from 'next/image'
-import Navigation from '@/components/layout/navigation'
+import Navigation from '@/components/common/layout/navigation'
 import styles from '../styles/Home.module.css'
 
 import vercelLogo from '@/public/vercel.svg'
